requirements.txt
pandas>=2.2.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
hmmlearn>=0.3.3,<1.0.0
matplotlib>=3.9.0,<4.0.0
SQLAlchemy>=2.0.0,<3.0.0
PyMySQL>=1.1.0,<2.0.0
requests>=2.32.0,<3.0.0
mysql-connector-python>=8.4.0,<10.0.0
python-dotenv>=1.0.1,<2.0.0
keyring>=25.0.0,<26.0.0
tensorflow>=2.16.0,<3.0.0
databento>=0.44.0,<1.0.0
